// src/adapters/types.ts
export type AdapterContractTypeLegacy =
  | "AD_AAVE_V3_LENDING_POOL"
  | "AD_BALANCER_VAULT"
  | "AD_CAMELOT_V3_ROUTER"
  | "AD_CONVEX_L2_BOOSTER"
  | "AD_CONVEX_L2_REWARD_POOL"
  | "AD_CONVEX_V1_BASE_REWARD_POOL"
  | "AD_CONVEX_V1_BOOSTER"
  | "AD_CURVE_STABLE_NG"
  | "AD_CURVE_V1_2ASSETS"
  | "AD_CURVE_V1_3ASSETS"
  | "AD_CURVE_V1_4ASSETS"
  | "AD_CURVE_V1_EXCHANGE_ONLY"
  | "AD_CURVE_V1_STECRV_POOL"
  | "AD_CURVE_V1_WRAPPER"
  | "AD_DAI_USDS_EXCHANGE"
  | "AD_EQUALIZER_ROUTER"
  | "AD_ERC4626_VAULT"
  | "AD_LIDO_V1"
  | "AD_LIDO_WSTETH_V1"
  | "AD_MELLOW_ERC4626_VAULT"
  | "AD_MELLOW_LRT_VAULT"
  | "AD_PENDLE_ROUTER"
  | "AD_STAKING_REWARDS"
  | "AD_SYMBIOTIC_DEFAULT_COLLATERAL"
  | "AD_UNISWAP_V2_ROUTER"
  | "AD_UNISWAP_V3_ROUTER"
  | "AD_VELODROME_V2_ROUTER"
  | "AD_YEARN_V2"
  | "AD_ZIRCUIT_POOL";

export type AdapterContractType =
  | "ADAPTER::AAVE_V3_LENDING_POOL"
  | "ADAPTER::BALANCER_V3_ROUTER"
  | "ADAPTER::BALANCER_VAULT"
  | "ADAPTER::CAMELOT_V3_ROUTER"
  | "ADAPTER::CONVEX_L2_BOOSTER"
  | "ADAPTER::CONVEX_L2_REWARD_POOL"
  | "ADAPTER::CURVE_STABLE_NG"
  | "ADAPTER::CURVE_V1_2ASSETS"
  | "ADAPTER::CURVE_V1_3ASSETS"
  | "ADAPTER::CURVE_V1_4ASSETS"
  | "ADAPTER::CURVE_V1_EXCHANGE_ONLY"
  | "ADAPTER::CURVE_V1_STECRV_POOL"
  | "ADAPTER::CURVE_V1_WRAPPER"
  | "ADAPTER::CVX_V1_BASE_REWARD_POOL"
  | "ADAPTER::CVX_V1_BOOSTER"
  | "ADAPTER::DAI_USDS_EXCHANGE"
  | "ADAPTER::EQUALIZER_ROUTER"
  | "ADAPTER::ERC4626_VAULT"
  | "ADAPTER::LIDO_V1"
  | "ADAPTER::LIDO_WSTETH_V1"
  | "ADAPTER::MELLOW_ERC4626_VAULT"
  | "ADAPTER::MELLOW_LRT_VAULT"
  | "ADAPTER::PENDLE_ROUTER"
  | "ADAPTER::STAKING_REWARDS"
  | "ADAPTER::SYMBIOTIC_DEFAULT_COLLATERAL"
  | "ADAPTER::UNISWAP_V2_ROUTER"
  | "ADAPTER::UNISWAP_V3_ROUTER"
  | "ADAPTER::VELODROME_V2_ROUTER"
  | "ADAPTER::YEARN_V2"
  | "ADAPTER::ZIRCUIT_POOL";